Namespace the Statsd class on the datadog module.
Otherwise you can't load it with any other type of statsd.
