+++ dogstatsd-ruby.gemspec
@@ -1,9 +1,9 @@
 $:.unshift File.expand_path("../lib", __FILE__)
-require "statsd"
+require "datadog/statsd"
 
 Gem::Specification.new do |s|
   s.name = "dogstatsd-ruby"
-  s.version = Statsd.VERSION
+  s.version = Datadog::Statsd.VERSION
 
   s.authors = ["Rein Henrichs"]
